Fixed Previous Scan function
