+++ sit/UI/HistoryView.swift
@@ -6,9 +6,12 @@
 //
 
 import SwiftUI
+import Firebase
 
 struct HistoryView: View {
     @EnvironmentObject var authState: AuthenticationState
+    
+    @State private var previousScans: [ScanResult] = []
     
     var body: some View {
         ZStack {
@@ -19,7 +22,29 @@ struct HistoryView: View {
                 Text("History")
                     .foregroundColor(.white)
             }
+        }.onAppear() {
+            getPreviousScans()
         }
+    }
+    
+    private func getPreviousScans() {
+        let db = Firestore.firestore()
+        
+        db.collection("scans")
+            .whereField("uid", isEqualTo: String(authState.user?.uid ?? ""))
+            .getDocuments() { (querySnapshot, error) in
+                if let error = error {
+                    //TODO: Handle error
+                } else {
+                    for document in querySnapshot!.documents {
+                        let data = document.data()
+                        let id = document.documentID
+                        if let scanResult = ScanResult(id: document.documentID, data: document.data()) {
+                            self.previousScans.append(scanResult)
+                        }
+                    }
+                }
+            }
     }
 }
